Added system update and checksum into UI

+++ system/system.py
@@ -54,12 +54,16 @@ def fileIntegrityRegister():
 	return checksum
 
 def getSystemFile(url,fileID,filename,fileHash):
+	filename = str(Path(filename))
+	filepath = str(Path(filename).parent)
+	if not jimi.helpers.safeFilepath(filename,filepath):
+		return False
 	headers = { "x-api-token" : jimi.auth.generateSystemSession() }
 	with requests.get("{0}{1}system/file/{2}/".format(url,jimi.api.base,fileID), headers=headers, stream=True, timeout=60) as r:
 		r.raise_for_status()
-		filename = str(Path(filename))
-		filepath = str(Path(filename).parent)
 		tempFilename = str(Path("data/temp/{0}".format(fileID)))
+		if not jimi.helpers.safeFilepath(tempFilename,"data/temp"):
+			return False
 		with open(tempFilename, 'wb') as f:
 			for chunk in r.iter_content(chunk_size=8192):
 				f.write(chunk)
@@ -69,7 +73,6 @@ def getSystemFile(url,fileID,filename,fileHash):
 			if not os.path.isdir(filepath):
 				os.makedirs(filepath)
 			os.rename(tempFilename,filename)
-			jimi.helpers.reload()
 			return True
 		else:
 			jimi.logging.debug("Error: File obtained from master failed integrity checks. fileID={0}, filename={1}".format(fileID,filename),-1)
@@ -84,7 +87,7 @@ def fixChecksum(pullFromSystemID):
 	for ourSystemFileHash in ourSystemFiles:
 		ourSystemFilesHash[str(Path(ourSystemFileHash.filename))] = ourSystemFileHash
 
-	masterURL = jimi.cluster.getMaster()
+	masterURL = jimi.cluster.getclusterMemberURLById(pullFromSystemID)
 	for masterSystemFile in masterSystemFiles:
 		try:
 			if ourSystemFilesHash[str(Path(masterSystemFile.filename))].fileHash != masterSystemFile.fileHash:
@@ -95,10 +98,11 @@ def fixChecksum(pullFromSystemID):
 			jimi.logging.debug("Info: File not present but is on master - Downloading. filename={0}".format(str(Path(masterSystemFile.filename))),-1)
 			getSystemFile(masterURL,masterSystemFile._id,str(Path(masterSystemFile.filename)),masterSystemFile.fileHash)
 	for ourSystemFile in ourSystemFilesHash:
-		jimi.logging.debug("Info: File does not exist on master - Deleting. filename={0}".format(str(Path(masterSystemFile.filename))),-1)
-		os.remove(ourSystemFile)
-	jimi.cluster.clusterMember.checksum = fileIntegrityRegister()
-	jimi.cluster.clusterMember.update(["checksum"])
+		jimi.logging.debug("Info: File does not exist on master - Deleting. filename={0}".format(str(Path(ourSystemFile))),-1)
+		if jimi.helpers.safeFilepath(ourSystemFile):
+			os.remove(ourSystemFile)
+	jimi.cluster.cluster.clusterMember.checksum = fileIntegrityRegister()
+	jimi.cluster.cluster.clusterMember.update(["checksum"])
 	return True
 
 # API
@@ -112,16 +116,40 @@ if jimi.api.webServer:
 				if len(systemFile) != 1:
 					return {}, 404
 				systemFile = systemFile[0]
+				if not jimi.helpers.safeFilepath(systemFile.filename):
+					return {}, 403
 				return jimi.api.send_file(systemFile.filename,attachment_filename=systemFile.name)
 
 			@jimi.api.webServer.route(jimi.api.base+"system/update/<pullFromSystemID>/", methods=["GET"])
 			@jimi.auth.adminEndpoint
 			def updateSystem(pullFromSystemID):
-				fixChecksum(pullFromSystemID)
+				fixChecksum(int(pullFromSystemID))
 				return {}, 200
 
 			@jimi.api.webServer.route(jimi.api.base+"system/checksum/", methods=["GET"])
 			@jimi.auth.adminEndpoint
 			def regenerateSystemFileIntegrity():
-				fileIntegrityRegister()
+				jimi.cluster.cluster.clusterMember.checksum = fileIntegrityRegister()
+				jimi.cluster.cluster.clusterMember.update(["checksum"])
 				return {}, 200
+		
+		if jimi.api.webServer.name == "jimi_web":
+			@jimi.api.webServer.route(jimi.api.base+"system/update/<systemID>/<pullFromSystemID>/", methods=["GET"])
+			@jimi.auth.adminEndpoint
+			def updateSystem(systemID,pullFromSystemID):
+				url = jimi.cluster.getclusterMemberURLById(int(systemID))
+				if not url:
+					return {}, 404
+				apiEndpoint = "system/update/{0}/".format(pullFromSystemID)
+				response = jimi.helpers.apiCall("GET",apiEndpoint,token=jimi.api.g.sessionToken,overrideURL=url)
+				return { "url" : url, "response" : response.status_code }, 200
+
+			@jimi.api.webServer.route(jimi.api.base+"system/checksum/<systemID>/", methods=["GET"])
+			@jimi.auth.adminEndpoint
+			def regenerateSystemFileIntegrity(systemID):
+				url = jimi.cluster.getclusterMemberURLById(int(systemID))
+				if not url:
+					return {}, 404
+				apiEndpoint = "system/checksum/"
+				response = jimi.helpers.apiCall("GET",apiEndpoint,token=jimi.api.g.sessionToken,overrideURL=url)
+				return { "url" : url, "response" : response.status_code }, 200
